requirements.txt
opencv-python>=4.8.0
numpy>=1.24.0
python-dotenv>=1.0.0
flask>=3.0.0
twilio>=8.10.0
schedule>=1.2.0
pillow>=10.0.0
python-dateutil>=2.8.2
